Implement OPA Generic Hibernate Dao
