+++ src/test/java/com/github/jferrater/opa/data/filter/spring/boot/starter/MyRepository.java
@@ -1,0 +1,13 @@
+package com.github.jferrater.opa.data.filter.spring.boot.starter;
+
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+@Repository
+@Transactional
+class MyRepository extends OpaGenericDataFilterDao<Pet> {
+
+    public MyRepository() {
+        setClazz(Pet.class);
+    }
+}
