+++ src/test/java/com/github/jferrater/opa/data/filter/spring/boot/starter/TestConfiguration.java
@@ -1,0 +1,21 @@
+package com.github.jferrater.opa.data.filter.spring.boot.starter;
+
+import com.zaxxer.hikari.HikariDataSource;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import javax.sql.DataSource;
+
+@Configuration
+public class TestConfiguration {
+
+    @Bean
+    public DataSource dataSource() {
+        final HikariDataSource dataSource = new HikariDataSource();
+        dataSource.setDriverClassName("org.h2.Driver");
+        dataSource.setJdbcUrl("jdbc:h2:mem:db;DB_CLOSE_DELAY=-1;INIT=RUNSCRIPT FROM 'classpath:sql/init.sql'");
+        dataSource.setUsername("sa");
+        dataSource.setPassword("");
+        return dataSource;
+    }
+}
